requirements.txt
aiosqlite==0.21.0
alembic==1.17.1
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
astroid==4.0.4
bcrypt==5.0.0
billiard==4.2.4
celery==5.6.0
certifi==2025.11.12
cffi==2.0.0
cfgv==3.5.0
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==46.0.3
-e git+https://github.com/Joaopydev/delivery-api.git@251fa1cd901a46c4f42942262f9bac6d8905dd70#egg=delivery_api
dill==0.4.1
distlib==0.4.0
distro==1.9.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
exceptiongroup==1.3.1
fastapi==0.121.1
filelock==3.25.1
flower==2.0.1
gevent==25.9.1
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
humanize==4.14.0
identify==2.6.17
idna==3.11
isort==8.0.1
Jinja2==3.1.6
jiter==0.12.0
kombu==5.6.1
mailtrap==2.4.0
Mako==1.3.10
MarkupSafe==3.0.3
mccabe==0.7.0
nodeenv==1.10.0
openai==2.14.0
packaging==25.0
platformdirs==4.9.4
pre_commit==4.5.1
prometheus_client==0.23.1
prompt_toolkit==3.0.52
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.4
pydantic-settings==2.12.0
pydantic_core==2.41.5
pylint==4.0.5
python-dateutil==2.9.0.post0
python-discovery==1.1.2
python-dotenv==1.2.1
python-http-client==3.3.7
python-jose==3.5.0
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.49.3
stripe==14.2.0
tomlkit==0.14.0
tornado==6.5.4
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.6.2
uvicorn==0.38.0
vine==5.1.0
virtualenv==21.2.0
wcwidth==0.2.14
Werkzeug==3.1.4
zope.event==6.1
zope.interface==8.1.1
